Added relevant comments to the code
- Applied code indentation so that it appears neat and aligned

- In "Factory contract Explained"-
a) Prefixed the parameters _salt and _bytecode with an underscore to follow the convention of using an underscore to indicate function parameters.
b) Changed the order of the parameters in the createContract function to match the order in which they are used in the function call.
c) Removed the unnecessary type conversion from the getAddress function.
d) Added error handling: In createContract, I have added a require statement to check if the contract was successfully deployed. If it fails, it reverts with an error message indicating the failure.

- I added the following error handling in " TestContract Explained":
a) In the constructor, added checks to ensure that the _owner address is not address(0) and that the _walletname string is not empty.
b) In the transferOwnership function, added a check to ensure that the _newAdmin address is not address(0) and that it is not the same as the current admin address.
c) In the withdraw function, added a check to ensure that the contract has a balance greater than zero.
d) In the receive function, added a check to ensure that the value sent with the transaction is greater than zero.